connection_scopes are ignored by A0WebViewAuthenticator
I am trying to use connection scopes with Facebook, but I am finding
they are ignored by A0WebViewAuthenticator - it seems that for other
IDPs the params get the 'connection' property set (in
registerAuthenticationProvider):

    params[A0ParameterConnection] = connectionName;

But this is not set for A0WebViewAuthenticator. Subsequently, when
A0WebKitViewController.initWithAPIClient is called, [parameters
asAPIPayload] returns a parameter dictionary with connection_scopes
missing, because it is trying to look up the connection scopes for the
selected connection, but no connection is set in the dictionary.

This is maybe more a band-aid than an actual fix for this issue, I'm not
sure, but it solves my problem.

see auth0/Lock.iOS-OSX#209