Fixes to passcode UI strings
The passcode was previously referred to, in this file, as passcode or email code.
The text placeholder "Email code" is confusing as "Email" can be interrupted as both a verb as well as a noun. The term "Enter" is clearer.
One text string is clipped on small-screen iPhones. I've shorten it so that it fits. The more concise text is easier for users to scan.